+++ dangerfile.ts
@@ -141,7 +141,7 @@ import { danger, fail } from "danger";
 
   function parseFirstCommitMessage(commitMessage: string) {
     const commitMessageRegex = new RegExp(
-      `^#(?<issueId>\\d+) \\[(?<type>${types.join("|")})\\] (?<scope>[^\\s].+[^\\s]): (?<summary>[^\\s].+[^\\s])`,
+      `^#(?<issueId>\\d+) \\[(?<type>${types.join("|")})\\] (?<scope>[^\\s][^:].+?[^\\s]): (?<summary>[^\\s].+[^\\s])`,
     );
 
     const match = commitMessage.match(commitMessageRegex);
@@ -168,7 +168,7 @@ import { danger, fail } from "danger";
 
   function parseChangelogEntry(changelog: string, issueId: number) {
     const changelogRegex = new RegExp(
-      `^\\* (?:\\*\\*(?<breaking>BREAKING)\\*\\* )?(?<scope>[^\\s].+[^\\s]): (?<summary>[^\\s].+[^\\s]) \\(\\[#${issueId}\\]\\(https:\\/\\/github\\.com\\/dso-toolkit\\/dso-toolkit\\/issues\\/(?<linkIssueId>\\d+)\\)\\)$`,
+      `^\\* (?:\\*\\*(?<breaking>BREAKING)\\*\\* )?(?<scope>[^\\s][^:].+?[^\\s]): (?<summary>[^\\s].+[^\\s]) \\(\\[#${issueId}\\]\\(https:\\/\\/github\\.com\\/dso-toolkit\\/dso-toolkit\\/issues\\/(?<linkIssueId>\\d+)\\)\\)$`,
       "m",
     );