+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * Document Component: Scrollen in tabel na vergroten werkt niet meer ([#2486](https://github.com/dso-toolkit/dso-toolkit/issues/2486))
 * Styling: Version number in dso.css `undefined` ([#2520](https://github.com/dso-toolkit/dso-toolkit/issues/2520))
 * Pagination: Long numbers overlap ([#2541](https://github.com/dso-toolkit/dso-toolkit/issues/2541))
+* Toggletip: Wordt niet getoond bij position: bottom ([#2477](https://github.com/dso-toolkit/dso-toolkit/issues/2477))
 
 ### Docs
 * Card: Link selecteren met href werkt niet ([#2506](https://github.com/dso-toolkit/dso-toolkit/issues/2506))